+++ org.hl7.fhir.utilities/src/main/java/org/hl7/fhir/utilities/tests/execution/CliTestException.java
@@ -1,7 +1,6 @@
 package org.hl7.fhir.utilities.tests.execution;
 
 public interface CliTestException {
-
-  public String getTestId();
-  public Throwable getException();
+  String getTestId();
+  Throwable getException();
 }
