+++ org.hl7.fhir.utilities/src/main/java/org/hl7/fhir/utilities/tests/execution/CliTestSummary.java
@@ -3,12 +3,11 @@ package org.hl7.fhir.utilities.tests.execution;
 import java.util.List;
 
 public interface CliTestSummary {
-  public long getTestsFoundCount();
-  public long getTestsFailedCount();
-  public long getTestsAbortedCount();
-  public long getTestsSkippedCount();
+  long getTestsFoundCount();
+  long getTestsFailedCount();
+  long getTestsAbortedCount();
+  long getTestsSkippedCount();
 
-  public List<CliTestException> getExceptions();
-
+  List<CliTestException> getExceptions();
 
 }
