--- org.hl7.fhir.utilities/src/main/java/org/hl7/fhir/utilities/tests/execution/TxCacheExtractor.java
@@ -1,8 +1,0 @@
-package org.hl7.fhir.utilities.tests.execution;
-
-public class TxCacheExtractor {
-
-  public static void copyTxCacheToDirectory(String cachePath, String targetDirectory) {
-
-  }
-}
